tui: Move file review action router
The TUI file review browser now keeps command execution helpers outside the
main loop, while the loop still owns standard line, file, and replacement
gating.

Those gates mix prompt collection, destructive confirmation, source-mode
checks, and transfer dispatch. That makes the browser harder to read than
the block and fixup paths that now delegate command work.

This commit addresses that by moving line, file, and replacement action
routing into file_review.action_router.

Subsequent commits will add an architecture guard for the router owner.
